+++ infrastructure-android/src/main/kotlin/extension/AndroidOptions.kt
@@ -1,26 +1,31 @@
 package com.redmadrobot.build.extension
 
-/** Settings for android modules. */
-public val RedmadrobotExtension.android: AndroidOptions by RedmadrobotExtension.field { AndroidOptions() }
+import org.gradle.api.provider.Property
+import org.gradle.kotlin.dsl.android as _android
 
 /** Settings for android modules. */
+@Suppress("DeprecatedCallableAddReplaceWith")
+@Deprecated("Property `android` can be used without import. Please remove import manually")
+public val RedmadrobotExtension.android: AndroidOptions
+    get() = _android
+
+/** Settings for android modules. */
+@Suppress("DeprecatedCallableAddReplaceWith")
+@Deprecated("Function `android { }` can be used without import. Please remove import manually")
 public fun RedmadrobotExtension.android(configure: AndroidOptions.() -> Unit) {
-    android.configure()
+    _android(configure)
 }
 
-public class AndroidOptions internal constructor() {
+public interface AndroidOptions : TestOptionsSpec {
 
     /** Minimal Android SDK that will be applied to all android modules. */
-    public var minSdk: Int = 21
+    public val minSdk: Property<Int>
 
     /** Target Android SDK that will be applied to all android modules. Also determines compile SDK. */
-    public var targetSdk: Int = 30
+    public val targetSdk: Property<Int>
 
-    /** Settings for Android test task. */
-    public val test: TestOptions = TestOptions()
-
-    /** Settings for Android test task. */
-    public fun test(configure: TestOptions.() -> Unit) {
-        test.run(configure)
+    public companion object {
+        internal const val DEFAULT_MIN_API = 21
+        internal const val DEFAULT_TARGET_API = 30
     }
 }