+++ infrastructure/src/main/kotlin/PublishPlugin.kt
@@ -1,5 +1,6 @@
 package com.redmadrobot.build
 
+import com.redmadrobot.build.extension.PublishingOptionsImpl
 import com.redmadrobot.build.extension.isReleaseVersion
 import com.redmadrobot.build.internal.java
 import org.gradle.api.Project
@@ -34,12 +35,12 @@ public open class PublishPlugin : InfrastructurePlugin() {
                 pom {
                     name.convention(project.name)
                     description.convention(project.description)
-                    options.configurePom(this)
+                    (options as PublishingOptionsImpl).configurePom.get().invoke(this)
                 }
             }
 
-            if (options.signArtifacts) {
-                configureSigning(options.useGpgAgent)
+            if (options.signArtifacts.get()) {
+                configureSigning(options.useGpgAgent.get())
             }
         }
     }