+++ infrastructure/src/main/kotlin/internal/Provider.kt
@@ -1,0 +1,8 @@
+package com.redmadrobot.build.internal
+
+import org.gradle.api.provider.Property
+
+internal fun <T : Any> Property<T>.setFinalValue(value: T) {
+    set(value)
+    finalizeValue()
+}