+++ infrastructure/src/main/kotlin/internal/Test.kt
@@ -1,18 +1,15 @@
 package com.redmadrobot.build.internal
 
 import com.redmadrobot.build.extension.TestOptions
+import com.redmadrobot.build.extension.TestOptionsImpl
 import org.gradle.api.tasks.testing.Test
 
 public fun Test.setTestOptions(testOptions: TestOptions) {
-    if (testOptions.useJunitPlatform) {
-        useJUnitPlatform {
-            excludeEngines = testOptions.jUnitPlatformOptions.excludeEngines
-            includeEngines = testOptions.jUnitPlatformOptions.includeEngines
-            excludeTags = testOptions.jUnitPlatformOptions.excludeTags
-            includeTags = testOptions.jUnitPlatformOptions.includeTags
-        }
+    val configure = (testOptions as TestOptionsImpl).configuration.get()
+    if (testOptions.useJunitPlatform.get()) {
+        useJUnitPlatform(configure)
         testLogging { events("passed", "skipped", "failed") }
     } else {
-        useJUnit()
+        useJUnit(configure)
     }
 }